fixed key

+++ hwpe-tb/sw/aes_hwpe.h
@@ -5,7 +5,7 @@
 
 void aes_hwpe_configure(uint8_t *input, uint8_t *output, uint32_t data_byte_length, uint32_t key_bit_length);
 
-void aes_hwpe_key_set(uint32_t *key_value);
+void aes_hwpe_key_set(uint8_t *key_value);
 
 void aes_hwpe_start(void);
